# openapi.yaml
...
openapi: 3.1.0
info:
  title: Nexus API
  version: 0.1.0
  description: "Generated from the live Fastify route table. Every path carries a
    maturity tag: `durable` — Dedicated handler over durable storage. Safe to
    build on. `dedicated-volatile` — Dedicated handler, but its state lives in
    process memory and is lost on restart. `bridge` — Served by the legacy
    in-memory bridge (routes/api-bridge.ts). Treat its data as synthetic."
  license:
    name: Apache-2.0
    url: https://www.apache.org/licenses/LICENSE-2.0
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
paths:
...
  /api/v1/connectors:
    get:
      tags:
        - durable
      responses:
        "200":
          description: Default Response
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
        "201":
          description: Default Response
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
  /api/v1/connectors/{id}:
    get:
      tags:
        - durable
      parameters:
        - schema:
            type: string
          in: path
          name: id
          required: true
      responses:
        "200":
          description: Default Response
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
        "201":
          description: Default Response
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
    patch:
      tags:
        - durable
      parameters:
        - schema:
            type: string
          in: path
          name: id
          required: true
      responses:
        "200":
          description: Default Response
  /api/v1/connectors/connect:
    post:
      tags:
        - durable
      responses:
        "200":
          description: Default Response
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
        "201":
          description: Default Response
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
...
  /api/v1/connectors/{id}/reconnect:
    post:
      tags:
        - durable
      parameters:
        - schema:
            type: string
          in: path
          name: id
          required: true
      responses:
        "200":
          description: Default Response
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
        "201":
          description: Default Response
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
  /api/v1/connectors/{id}/disconnect:
    post:
      tags:
        - durable
      parameters:
        - schema:
            type: string
          in: path
          name: id
          required: true
      responses:
        "200":
          description: Default Response
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
        "201":
          description: Default Response
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
  /api/v1/connectors/{id}/oauth/start:
    get:
      tags:
        - durable
      parameters:
        - schema:
            type: string
          in: path
          name: id
          required: true
      responses:
        "200":
          description: Default Response
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
        "201":
          description: Default Response
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
  /api/v1/connectors/{id}/oauth/callback:
    get:
      tags:
        - durable
      parameters:
        - schema:
            type: string
          in: path
          name: id
          required: true
      responses:
        "200":
          description: Default Response
  /api/v1/connectors/{id}/oauth/credential:
    get:
      tags:
        - durable
      parameters:
        - schema:
            type: string
          in: path
          name: id
          required: true
      responses:
        "200":
          description: Default Response
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
        "404":
          description: Default Response
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
...
  /api/connectors/{id}:
    delete:
      tags:
        - dedicated-volatile
      parameters:
        - schema:
            type: string
          in: path
          name: id
          required: true
      responses:
        "200":
          description: Default Response
...
security:
  - bearerAuth: []
...